Updated Dependencies, Improved Initial Loadtime, Higher Resolution Thumbnails
This is the first push with any meaningful changes versus the initial fork and includes the following:

• All standard dependencies are updated to the latest version. devDependencies will come later. This should fix a few issues relating to memory leaks and anything else that has been improved since these were last bumped.

The load-time has been reduced to around 5 seconds, versus the previous 15+ seconds. This was primarily done by lowering a sleep/pause to give everything time to authenticate, but it seems to do fine with the lower number. This may change in the future.

Thumbnails for videos are now twice the resolution they were before.