docs: update README to reflect fork, Roku Store
Previously the README referenced the old repo, this updates the relevant references to the forked
repo. In addition, the README has been updated with a corrected link to the most recent release
download where before there was a typo anyway. Lastly the README has been updated to clarify that
installation is still possible, just not from the Roku Store directly.